+++ mining/cpuminer/cpuminer.go
@@ -206,7 +206,7 @@ func (m *CPUMiner) submitBlock(block *util.Block) bool {
 // This function will return early with false when conditions that trigger a
 // stale block such as a new block showing up or periodically when there are
 // new transactions and enough time has elapsed without finding a solution.
-func (m *CPUMiner) solveBlock(msgBlock *wire.MsgBlock, blockHeight int32,
+func (m *CPUMiner) solveBlock(msgBlock *wire.MsgBlock, blockHeight uint64,
 	ticker *time.Ticker, quit chan struct{}) bool {
 
 	// Create some convenience variables.
